requirements.txt
flask
Flask==2.3.3
Pillow==10.0.0
rembg==2.0.50
requests==2.31.0